신종하 : 경고제거

+++ GameEngineBase/GameEngineTime.cpp
@@ -29,7 +29,7 @@ void GameEngineTime::TimeCheckReset()
 void GameEngineTime::TimeCheck()
 {
 	QueryPerformanceCounter(&endCheck_);
-	deltaTime_ = static_cast<double>((endCheck_.QuadPart - startCheck_.QuadPart)) / static_cast<double>(timeCount_.QuadPart);
+	deltaTime_ = static_cast<float>(static_cast<double>((endCheck_.QuadPart - startCheck_.QuadPart)) / static_cast<double>(timeCount_.QuadPart));
 	startCheck_.QuadPart = endCheck_.QuadPart;
 
 	if (deltaTime_ > 0.01666f)
